Move components to be shared
